+++ lib/dialog/beer_consumption_options_dialog.dart
@@ -3,6 +3,7 @@
 
 import "package:flutter/material.dart";
 import "package:my_beer_diary/common.dart";
+import "package:my_beer_diary/dialog/app_alert_dialog.dart";
 import "package:my_beer_diary/dialog/beer_consumption_dialog.dart";
 import "package:my_beer_diary/dialog/beer_consumption_move_dialog.dart";
 import "package:my_beer_diary/dialog/beer_dialog.dart";
@@ -95,23 +96,14 @@ class BeerConsumptionOptionsDialog extends StatelessWidget {
 
                         final result = await showDialog(
                           context: context,
-                          builder: (BuildContext context) => AlertDialog(
-                            title: Text("Smazat záznam"),
-                            content: Text(
-                              "Opravdu si přejete smazat vypití piva „${beer.toDisplayString()}“?",
-                            ),
-                            actions: [
-                              TextButton(
-                                onPressed: () =>
-                                    Navigator.of(context).pop(false),
-                                child: Text("Zrušit"),
-                              ),
-                              TextButton(
-                                onPressed: () =>
-                                    Navigator.of(context).pop(true),
-                                child: Text("Smazat"),
-                              ),
-                            ],
+                          builder: (BuildContext context) => AppAlertDialog(
+                            titleText: "Smazat záznam",
+                            bodyText:
+                                "Opravdu si přejete smazat vypití piva „${beer.toDisplayString()}“?",
+
+                            yesText: "Smazat",
+                            noText: "Zrušit",
+                            doShowNo: true,
                           ),
                         );