+++ asset/markdown/changelog.md
@@ -5,3 +5,7 @@
 ### verze 0.1.0 (14.09.2026 19:43)
 
 * První vydání
+
+## Sestavení
+
+`flutter build apk`