+++ asset/markdown/changelog.md
@@ -2,6 +2,10 @@
 
 ## Changelog
 
+### verze 0.2.0 (WIP)
+
+* Přidána možnost zálohy databáze do souboru a obnovení zálohy ze souboru, např. na novém zařízení
+
 ### verze 0.1.0 (14.09.2026 19:43)
 
 * První vydání